Util tweaks
- Move rand util into rand.go
- RandString changed to RandomID and uses 256 bits
- Move string util into strings.go
